fix(auth): 登录令牌过期时间以数值返回
expiresIn/refreshExpiresIn 语义为秒数而非雪花 ID，改用 int 输出 JSON number，避免被全局 Long→String 序列化错误地输出为字符串，与前端 schema（期望正整数）一致。

+++ src/main/java/com/leo/erp/auth/service/TokenIssuanceService.java
@@ -133,13 +133,13 @@ public class TokenIssuanceService {
 
         eventPublisher.publishEvent(new SessionInvalidatedEvent(user.getId(), sessionTokenId, false));
 
-        long refreshExpiresIn = jwtTokenService.getRefreshExpirationMs() / MILLIS_PER_SECOND;
+        int refreshExpiresIn = (int) (jwtTokenService.getRefreshExpirationMs() / MILLIS_PER_SECOND);
 
         return new TokenResponse(
                 accessToken,
                 rawRefreshToken,
                 TOKEN_TYPE_BEARER,
-                jwtTokenService.getAccessExpirationMs() / MILLIS_PER_SECOND,
+                (int) (jwtTokenService.getAccessExpirationMs() / MILLIS_PER_SECOND),
                 refreshExpiresIn,
                 new AuthUserResponse(
                         user.getId(),
@@ -159,13 +159,13 @@ public class TokenIssuanceService {
         );
         String accessToken = jwtTokenService.generateAccessToken(principal, session.getTokenId());
 
-        long refreshExpiresIn = jwtTokenService.getRefreshExpirationMs() / MILLIS_PER_SECOND;
+        int refreshExpiresIn = (int) (jwtTokenService.getRefreshExpirationMs() / MILLIS_PER_SECOND);
 
         return new TokenResponse(
                 accessToken,
                 rawRefreshToken,
                 TOKEN_TYPE_BEARER,
-                jwtTokenService.getAccessExpirationMs() / MILLIS_PER_SECOND,
+                (int) (jwtTokenService.getAccessExpirationMs() / MILLIS_PER_SECOND),
                 refreshExpiresIn,
                 new AuthUserResponse(
                         user.getId(),
